requirements.txt
streamlit
pdfplumber
langchain
langchain-core
langchain-community
langchain-openai
langchain-ollama
chromadb==0.4.22
python-dotenv
openai
azure-identity
azure-core
unstructured[pdf]
pillow-heif
pillow
pdf2image
pypdf
pytesseract
unstructured_pytesseract
pi_heif
unstructured_inference